Run Cox Analysis in Batch Mode

Run Cox Analysis in Batch Mode

ezcox(data, covariates, controls = NULL, time = "time",
  status = "status", global_method = c("likelihood", "wald",
  "logrank"), return_models = FALSE, parallel = FALSE)

Arguments

data
a data.frame.

covariates
column names specifying variables.

controls
column names specifying controls.

time
column name specifying time, default is 'time'.

status
column name specifying event status, default is 'status'.

global_method
method used to obtain global p value for cox model, should be one of "likelihood", "wald", "logrank". The likelihood-ratio test, Wald test, and score logrank statistics. These three methods are asymptotically equivalent. For large enough N, they will give similar results. For small N, they may differ somewhat. The Likelihood ratio test has better behavior for small sample sizes, so it is generally preferred.

return_models
default FALSE. If TRUE, return a list contains cox models.

parallel
if TRUE, do parallel computation by furrr package.

Value

a tibble or a list

Examples

library(survival)

# Build unvariable models
ezcox(lung, covariates = c("age", "sex", "ph.ecog"))

#> => Processing variable age

#> ==> Building Surv object...

#> ==> Building Cox model...

#> ==> Done.

#> => Processing variable sex

#> ==> Building Surv object...

#> ==> Building Cox model...

#> ==> Done.

#> => Processing variable ph.ecog

#> ==> Building Surv object...

#> ==> Building Cox model...

#> ==> Done.

#> # A tibble: 3 x 11
#>   Variable contrast_level ref_level n_contrast n_ref    beta    HR lower_95
#>   <chr>    <chr>          <chr>          <int> <int>   <dbl> <dbl>    <dbl>
#> 1 age      age            age              228   228  0.0187 1.02     1    
#> 2 sex      sex            sex              228   228 -0.531  0.588    0.424
#> 3 ph.ecog  ph.ecog        ph.ecog          227   227  0.476  1.61     1.29 
#> # … with 3 more variables: upper_95 <dbl>, p.value <dbl>, global.pval <dbl>


# Build multi-variable models
# Control variable 'age'
ezcox(lung, covariates = c("sex", "ph.ecog"), controls = "age")

#> => Processing variable sex

#> ==> Building Surv object...

#> ==> Building Cox model...

#> ==> Done.

#> => Processing variable ph.ecog

#> ==> Building Surv object...

#> ==> Building Cox model...

#> ==> Done.

#> # A tibble: 4 x 11
#>   Variable contrast_level ref_level n_contrast n_ref    beta    HR lower_95
#>   <chr>    <chr>          <chr>          <int> <int>   <dbl> <dbl>    <dbl>
#> 1 sex      sex            sex              228   228 -0.513  0.599    0.431
#> 2 sex      age            age              228   228  0.017  1.02     0.999
#> 3 ph.ecog  ph.ecog        ph.ecog          227   227  0.443  1.56     1.24 
#> 4 ph.ecog  age            age              228   228  0.0113 1.01     0.993
#> # … with 3 more variables: upper_95 <dbl>, p.value <dbl>, global.pval <dbl>


# Return models
ezcox(lung, covariates = c("age", "sex", "ph.ecog"),
            return_models = TRUE)

#> => Processing variable age

#> ==> Building Surv object...

#> ==> Building Cox model...

#> ==> Done.

#> => Processing variable sex

#> ==> Building Surv object...

#> ==> Building Cox model...

#> ==> Done.

#> => Processing variable ph.ecog

#> ==> Building Surv object...

#> ==> Building Cox model...

#> ==> Done.

#> $res
#> # A tibble: 3 x 11
#>   Variable contrast_level ref_level n_contrast n_ref    beta    HR lower_95
#>   <chr>    <chr>          <chr>          <int> <int>   <dbl> <dbl>    <dbl>
#> 1 age      age            age              228   228  0.0187 1.02     1    
#> 2 sex      sex            sex              228   228 -0.531  0.588    0.424
#> 3 ph.ecog  ph.ecog        ph.ecog          227   227  0.476  1.61     1.29 
#> # … with 3 more variables: upper_95 <dbl>, p.value <dbl>, global.pval <dbl>
#> 
#> $models
#> # A tibble: 3 x 4
#>   Variable control model   status
#>   <chr>    <chr>   <list>  <lgl> 
#> 1 age      <NA>    <coxph> TRUE  
#> 2 sex      <NA>    <coxph> TRUE  
#> 3 ph.ecog  <NA>    <coxph> TRUE  
#> 
#> attr(,"class")
#> [1] "ezcox"

ezcox(lung, covariates = c("sex", "ph.ecog"), controls = "age",
            return_models = TRUE)

#> => Processing variable sex

#> ==> Building Surv object...

#> ==> Building Cox model...

#> ==> Done.

#> => Processing variable ph.ecog

#> ==> Building Surv object...

#> ==> Building Cox model...

#> ==> Done.

#> $res
#> # A tibble: 4 x 11
#>   Variable contrast_level ref_level n_contrast n_ref    beta    HR lower_95
#>   <chr>    <chr>          <chr>          <int> <int>   <dbl> <dbl>    <dbl>
#> 1 sex      sex            sex              228   228 -0.513  0.599    0.431
#> 2 sex      age            age              228   228  0.017  1.02     0.999
#> 3 ph.ecog  ph.ecog        ph.ecog          227   227  0.443  1.56     1.24 
#> 4 ph.ecog  age            age              228   228  0.0113 1.01     0.993
#> # … with 3 more variables: upper_95 <dbl>, p.value <dbl>, global.pval <dbl>
#> 
#> $models
#> # A tibble: 2 x 4
#>   Variable control model   status
#>   <chr>    <chr>   <list>  <lgl> 
#> 1 sex      age     <coxph> TRUE  
#> 2 ph.ecog  age     <coxph> TRUE  
#> 
#> attr(,"class")
#> [1] "ezcox"